+++ contracts/finality/src/state/finality.rs
@@ -32,12 +32,12 @@ pub struct Evidence {
 }
 
 /// Map of signatures by block height and fp
-pub(crate) const SIGNATURES: Map<(u64, &str), Vec<u8>> = Map::new("fp_sigs");
+pub(crate) const SIGNATURES: Map<(u64, &str), Vec<u8>> = Map::new("signatures");
 
 /// Map of block hashes by block height and fp
 pub(crate) const BLOCK_HASHES: Map<(u64, &str), Vec<u8>> = Map::new("block_hashes");
 
-/// Map of (block height, block hash) tuples to the list of fps that voted for this combination
+/// Map of (block height, block hash) tuples to the list of BTC PKs (in hex) of finality providers that voted for this combination
 pub(crate) const BLOCK_VOTES: Map<(u64, &[u8]), HashSet<String>> = Map::new("block_votes");
 
 /// Map of evidence by block height and fp